# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fleet Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /vehicles:
    get:
      operationId: listVehicles
      tags: [vehicles]
      summary: List all vehicles
      responses:
        "200":
          description: List of vehicles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vehicle"
    post:
      operationId: createVehicle
      tags: [vehicles]
      summary: Create a new vehicle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVehicleRequest"
      responses:
        "201":
          description: Vehicle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get vehicle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    put:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVehicleRequest"
      responses:
        "200":
          description: Vehicle updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    delete:
      operationId: deleteVehicle
      tags: [vehicles]
      summary: Delete vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [admin, driver]
        active:
          type: boolean
        vehicleId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required: [id, name, phone, role, active, createdAt]
    CreateUserRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        pin:
          type: string
        role:
          type: string
          enum: [admin, driver]
        vehicleId:
          type: integer
          nullable: true
      required: [name, phone, pin, role]
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        pin:
          type: string
        role:
          type: string
          enum: [admin, driver]
        vehicleId:
          type: integer
          nullable: true
        active:
          type: boolean
    Vehicle:
      type: object
      properties:
        id:
          type: integer
        plate:
          type: string
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        status:
          type: string
          enum: [active, maintenance, inactive]
        mileage:
          type: number
        fuelType:
          type: string
        assignedDriverId:
          type: integer
          nullable: true
        assignedDriverName:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, plate, brand, model, year, status, mileage, fuelType, createdAt]
    CreateVehicleRequest:
      type: object
      properties:
        plate:
          type: string
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        status:
          type: string
          enum: [active, maintenance, inactive]
        mileage:
          type: number
        fuelType:
          type: string
        assignedDriverId:
          type: integer
          nullable: true
      required: [plate, brand, model, year, status, mileage, fuelType]
    UpdateVehicleRequest:
      type: object
      properties:
        plate:
          type: string
        brand:
          type: string
        model:
          type: string
        year:
          type: integer
        status:
          type: string
          enum: [active, maintenance, inactive]
        mileage:
          type: number
        fuelType:
          type: string
        assignedDriverId:
          type: integer
          nullable: true
...